server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
qdrant-client
sqlalchemy
psycopg2-binary
asyncpg
sentence-transformers
google-generativeai
langchain
sentence-transformers
unstructured[md]
google-genai
pytest
pytest-asyncio
